LayoutTests/http/tests/workers/service/shownotification-allowed-expected.txt

This tests that a notification shown by a service worker can be clicked and closed, with those actions being observable by the service worker

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "granted"
PASS gotClicked is false
PASS gotClosed is false
PASS gotClicked is true
PASS gotClosed is false
PASS Close has been observed
